# Conflicts:
#	src/server/Organization/Controllers/ProjectController.cs
#	src/server/Organization/Models/Organization/Project/Board.cs

+++ src/server/Organization/Models/Organization/Accounts/UserGroup.cs
@@ -52,7 +52,7 @@ public class UserGroup : IUserAggregator
     /// Локальный идентификатор группы пользователей
     /// </summary>
     [Column("local_id")]
-    public required ulong LocalId { get; init; }
+    public required short LocalId { get; init; }
     
     /// <summary>
     /// Название группы пользователей
